Delete all per-technology / partial config plugins (#206)
We believe that per-technology and partial config plugins are the wrong
abstraction for sewing-kit-next plugins. For the best developer
understandability and ease of modification one plugin should offer an
opinionated configuration of a tool, rather than spreading that
configuration across multiple packages.

Having this layer that you could supposedly pick and choose from is
unneeded complexity and overabstraction that makes code harder to reason
about and modify.

The following packages have been removed:

- graphql
- plugin-css
- plugin-differential-serving
- plugin-graphql
- plugin-flexible-outputs
- plugin-react
- plugin-sass
- plugin-service-base
- plugin-vscode
- plugin-web-app-base
- plugin-webpack-builds

I don't believe the code in these packages will be useful in the future
as we shall need to modify it all for webpack 5 anyway.