Fix entrypoint and binfile generation for non-src directories
Previously entry and binfile generation assumed that source code lived
in the src directory. However rollup finds the common root of the
entries you specify and uses that as the base.

Now we find the common base and use that for generating import paths,
this means that we support both

- All source files are in a folder not called src
- Source files are spread across multiple folders within your project
  (this is unlikely but rollup handles it so we should too)